refactor: Merge the two variants of map saving forms
BREAKING CHANGE: advancedForm hsConfig parameter removed
