+++ projects/hslayers/src/assets/locales/cs.json
@@ -250,7 +250,8 @@
     "copyLayer": "Kopírovat vrstvu",
     "selectRecordCountPerPage": "Vyberte počet záznamů na stránku",
     "unknown": "Neznámý",
-    "selectAndConfirmToDeleteMultiple": "Vybrat a potvrdit smazání více vrstev"
+    "selectAndConfirmToDeleteMultiple": "Vybrat a potvrdit smazání více vrstev",
+    "aboutAuthor": "O autorovi"
   },
   "COMPOSITIONS": {
     "addByAddress": "Přidat kompozici z adresy",
