+++ projects/hslayers/src/components/save-map/partials/panel.html
@@ -15,9 +15,9 @@
             <div class="alert alert-info text-gray" translate>
                 {{'SAVECOMPOSITION.panelMd.youAreNotAuthorized' | translate}} </div>
         </div>
-        <hs-layman-current-user *ngIf="endpoint?.type === 'layman'" [endpoint]="endpoint"></hs-layman-current-user>
-        <hs-save-map-advanced-form class="d-block pt-5" *ngIf="advancedForm && (isAuthorized || endpoint?.type !== 'layman')">
-        </hs-save-map-advanced-form>
-        <hs-save-map-simple-form *ngIf="!advancedForm"></hs-save-map-simple-form>
+        <hs-layman-current-user class="d-block pb-5" *ngIf="endpoint?.type === 'layman'" [endpoint]="endpoint">
+        </hs-layman-current-user>
+        <hs-save-map-form *ngIf="(isAuthorized || endpoint?.type !== 'layman')">
+        </hs-save-map-form>
     </div>
 </div>
