--- projects/hslayers/src/components/save-map/partials/simpleform.html
@@ -1,87 +1,0 @@
-<div *ngIf="isAllowed()">
-    <form class="form">
-        <label class="form-group">{{'COMMON.name' | translate}}
-            <input type="text" class="form-control" [placeholder]="'COMMON.fillInName' | translate"
-                [(ngModel)]="HsSaveMapManagerService.compoData.title">
-        </label>
-
-        <label class="form-group">{{'COMMON.abstract' | translate}}
-            <textarea class="form-control" [placeholder]="'COMMON.fillInDescriptive' | translate"
-                [(ngModel)]="HsSaveMapManagerService.compoData.abstract" [ngModelOptions]="{standalone: true}">            </textarea>
-        </label>
-
-        <label class="form-group">{{'COMMON.thumbnail' | translate}}
-            <img class="hs-stc-thumbnail img-thumbnail" alt="Composition thumbnail" width="256" height="256">
-        </label>
-
-        <div class="form-group">
-            <div class="checkbox" *ngFor="let g of HsSaveMapManagerService.statusData.groups">
-                <label>
-                    <input type="checkbox" [(ngModel)]="g.r" [ngModelOptions]="{standalone: true}">
-                    <span>{{'SAVECOMPOSITION.panelMd.makePublic' | translate}}</span>
-                </label>
-            </div>
-        </div>
-
-    </form>
-
-    <button type="button" class="btn btn-secondary" data-toggle="collapse"
-        data-target="hs-save-comp-advanced-panel">{{'COMMON.advancedOptions' | translate}}</button>
-    <div class="collapse hs-save-comp-advanced-panel">
-        <form class="form">
-            <label class="form-group">{{'COMMON.keywords' | translate}}
-                <input type="text" class="form-control" [(ngModel)]="HsSaveMapManagerService.compoData.keywords" [ngModelOptions]="{standalone: true}">
-            </label>
-
-            <label class="form-group">{{'COMMON.extent' | translate}}
-                <div class="row hs-comp-ext-row">
-                    <div class="form-group hs-comp-ext col-sm-3">
-                        <input type="text" class="form-control input-group-lg"
-                            [(ngModel)]="HsSaveMapManagerService.compoData.bbox[0]" [ngModelOptions]="{standalone: true}">
-                    </div>
-                    <div class="form-group hs-comp-ext col-sm-3">
-                        <input type="text" class="form-control hs-comp-ext input-group-lg"
-                            [(ngModel)]="HsSaveMapManagerService.compoData.bbox[1]" [ngModelOptions]="{standalone: true}">
-                    </div>
-                    <div class="form-group hs-comp-ext col-sm-3">
-                        <input type="text" class="form-control hs-comp-ext input-group-lg"
-                            [(ngModel)]="HsSaveMapManagerService.compoData.bbox[2]" [ngModelOptions]="{standalone: true}">
-                    </div>
-                    <div class="form-group hs-comp-ext col-sm-3">
-                        <input type="text" class="form-control hs-comp-ext input-group-lg"
-                            [(ngModel)]="HsSaveMapManagerService.compoData.bbox[3]" [ngModelOptions]="{standalone: true}">
-                    </div>
-                </div>
-            </label>
-
-            <button type="button" class="btn btn-secondary"
-                (click)="HsSaveMapManagerService.setCurrentBoundingBox()">{{'SAVECOMPOSITION.form.userCurrentMapExtent' | translate}}</button>
-
-
-            <label class="form-group">{{'COMMON.layers' | translate}}
-                <ul class="list-group">
-                    <li *ngFor="let layer of HsSaveMapManagerService.compoData.layers; let index = index" id="layer-{{index}}"
-                        class="list-group-item">
-                        <span>{{HsLayerUtilsService.translateTitle(layer.title)}}</span>
-                        <span class="pull-right">
-                            <input type="checkbox" [(ngModel)]="layer.checked" [ngModelOptions]="{standalone: true}">
-                        </span>
-                    </li>
-                </ul>
-            </label>
-
-        </form>
-    </div>
-
-    <div class="card-footer row hs-stc-footer">
-        <div class="btn-group" role="group" [attr.aria-label]="'SAVECOMPOSITION.dialogSave.saveComposition' | translate">
-                <button type="button" class="btn btn-primary"
-                [disabled]="!HsSaveMapManagerService.compoData.title"
-                (click)="HsSaveMapManagerService.confirmSave()">{{'COMMON.save' | translate}}</button>
-        </div>
-    </div>
-</div>
-<div *ngIf="isAllowed()">
-    <div class="alert alert-info">
-        {{'SAVECOMPOSITION.panelMd.youAreNotAuthorized' | translate}}</div>
-</div>
