+++ projects/hslayers/src/components/save-map/public-api.ts
@@ -3,10 +3,9 @@ export * from './layer-synchronizer.service';
 export * from './layman-layer-descriptor.interface';
 export * from './layman-utils';
 export * from './layman.service';
-export * from './save-map-advanced-form.component';
+export * from './save-map-form.component';
 export * from './save-map-dialog.component';
 export * from './save-map-manager.service';
-export * from './save-map-simple-form.component';
 export * from './save-map.component';
 export * from './save-map.module';
 export * from './save-map.result-dialog.component';
