+++ projects/hslayers/src/components/save-map/save-map-form.component.ts
@@ -4,62 +4,36 @@ import {Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
 
 import {HsCoreService} from '../core/core.service';
-import {HsEventBusService} from '../core/event-bus.service';
 import {HsLayerUtilsService} from '../utils/layer-utils.service';
 import {HsSaveMapManagerService} from './save-map-manager.service';
-import {HsUtilsService} from './../utils/utils.service';
+import {HsUtilsService} from '../utils/utils.service';
+
 @Component({
-  selector: 'hs-save-map-advanced-form',
+  selector: 'hs-save-map-form',
   templateUrl: './partials/form.html',
 })
 export class HsSaveMapAdvancedFormComponent implements OnDestroy {
   btnSelectDeselectClicked = true;
-  step = 'context';
-  steps = ['context', 'access', 'author'];
   endpoint: any;
   overwrite = false;
   downloadableData: string;
+  extraFormOpened = '';
   private ngUnsubscribe = new Subject<void>();
   constructor(
     public HsSaveMapManagerService: HsSaveMapManagerService,
-    public HsEventBusService: HsEventBusService,
     public HsCoreService: HsCoreService,
     public HsUtilsService: HsUtilsService,
     public HsLayerUtilsService: HsLayerUtilsService //Used in template
   ) {
-    this.HsEventBusService.mapResets
-      .pipe(takeUntil(this.ngUnsubscribe))
-      .subscribe(() => {
-        this.step = 'context';
-      });
-
-    this.HsEventBusService.mainPanelChanges
-      .pipe(takeUntil(this.ngUnsubscribe))
-      .subscribe((which: string) => {
-        if (which == 'saveMap') {
-          this.step = 'context';
-        }
-      });
-
     this.HsSaveMapManagerService.endpointSelected
       .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe((endpoint) => {
         this.endpoint = endpoint;
-        switch (endpoint?.type) {
-          case 'layman':
-            this.steps = ['context', 'author'];
-            break;
-          default:
-            this.steps = ['context', 'access', 'author'];
-        }
       });
 
     this.HsSaveMapManagerService.saveMapResulted
       .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe((statusData) => {
-        if (statusData.status || statusData == 'rename') {
-          this.step = 'context';
-        }
         if (statusData.overWriteNeeded) {
           this.overwrite = true;
         }
@@ -68,6 +42,14 @@ export class HsSaveMapAdvancedFormComponent implements OnDestroy {
   ngOnDestroy(): void {
     this.ngUnsubscribe.next();
     this.ngUnsubscribe.complete();
+  }
+
+  setExtraFormTo(form: string): void {
+    if (this.extraFormOpened === form) {
+      this.extraFormOpened = '';
+    } else {
+      this.extraFormOpened = form;
+    }
   }
 
   saveCompoJson(): void {
@@ -93,24 +75,11 @@ export class HsSaveMapAdvancedFormComponent implements OnDestroy {
     );
   }
 
-  /**
-   * Callback function for clicking Next button, create download link for map context and show save, saveas buttons
-   */
-  next() {
-    const ixCurrent = this.steps.indexOf(this.step);
-    if (this.steps.length > ixCurrent + 1) {
-      this.step = this.steps[ixCurrent + 1];
-    } else {
-      this.step = 'end';
-    }
-  }
-
-  setStep(step) {
-    this.step = step;
-  }
+  //*NOTE not being used
   capitalizeFirstLetter(string: string): string {
     return this.HsUtilsService.capitalizeFirstLetter(string);
   }
+
   titleChanged() {
     this.overwrite = false;
   }
